requirements/base.txt
alembic==1.13.2
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
asyncpg==0.29.0
attrs==22.2.0
black==24.4.2
build==0.10.0
CacheControl==0.12.11
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cleo==2.0.1
click==8.1.7
crashtest==0.4.1
cryptography==39.0.2
distlib==0.3.6
dnspython==2.6.1
dulwich==0.21.3
email_validator==2.2.0
fastapi==0.111.0
fastapi-cli==0.0.4
filelock==3.9.0
greenlet==3.0.3
h11==0.14.0
html5lib==1.1
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.4
importlib-metadata==6.0.0
installer==0.6.0
jaraco.classes==3.2.3
jeepney==0.8.0
Jinja2==3.1.4
jsonschema==4.17.3
keyring==23.13.1
lockfile==0.12.2
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
more-itertools==9.1.0
msgpack==1.0.5
mypy-extensions==1.0.0
orjson==3.10.6
packaging==23.0
pathspec==0.12.1
pexpect==4.8.0
pipenv==2023.2.18
pkginfo==1.9.6
platformdirs==2.6.2
poetry==1.4.0
poetry-core==1.5.1
poetry-plugin-export==1.3.0
ptyprocess==0.7.0
pycparser==2.21
pydantic==2.8.2
pydantic-settings==2.3.4
pydantic_core==2.20.1
Pygments==2.18.0
PyJWT==2.9.0
pyproject_hooks==1.0.0
pyrsistent==0.19.3
pytest==8.3.2
pytest-asyncio==0.23.8
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
rapidfuzz==2.15.1
requests==2.28.2
requests-toolbelt==0.10.1
rich==13.7.1
ruff==0.5.1
SecretStorage==3.3.3
shellingham==1.5.0.post1
sniffio==1.3.1
SQLAlchemy==2.0.31
starlette==0.37.2
tomlkit==0.11.6
trove-classifiers==2023.3.9
typer==0.12.3
typing_extensions==4.12.2
ujson==5.10.0
urllib3==1.26.15
uvicorn==0.30.1
virtualenv==20.20.0
virtualenv-clone==0.5.7
watchfiles==0.22.0
webencodings==0.5.1
websockets==12.0
zipp==3.15.0
